refactor(project): Restructure core project files and update documentation
- Move main project files into core/ directory
- Rename and relocate generate_reports.py, main.py, and restructure_questions.py
- Add core/__init__.py to define core scraper functionality
- Update CONTRIBUTING.md with new project structure
- Create PROJECT_STRUCTURE.md documentation file
- Remove standalone image_downloader.py
- Update import paths in restructure_questions.py
- Improve project organization and modularity
This restructuring enhances code organization, improves import management, and provides a clearer project layout for better maintainability.
