+++ core/__init__.py
@@ -1,0 +1,1 @@
+# Core scraper functionality for BECE Questions Data Scraper
